requirements.txt
alembic==1.0.9
astroid==2.2.5
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.4.1
decorator==4.4.0
Flask==1.1.2
Flask-Login==0.4.1
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.3.2
Flask-SSLify==0.1.5
Flask-WTF==0.14.3
future==0.17.1
geocoder==1.38.1
gunicorn==20.0.4
idna==2.8
isort==4.3.17
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.3.1
Mako==1.0.9
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.0.1
psycopg2-binary==2.8.5
pylint==2.3.1
PyMySQL==0.9.3
pyTelegramBotAPI==3.6.6
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
pytz==2019.1
ratelim==0.1.6
requests==2.24.0
requests-oauthlib==1.2.0
six==1.12.0
SQLAlchemy==1.3.3
sqlalchemy-mptt==0.2.5
typed-ast==1.4.1
urllib3==1.24.1
Werkzeug==1.0.1
wrapt==1.11.1
WTForms==2.2.1
xlrd==1.2.0
yandex-geocoder==2.0.0
